requirements.txt
django
django-crispy-forms
crispy-bootstrap4
gunicorn
psycopg2-binary
whitenoise
Pillow
python-decouple
